+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
  - `iMethod` => `vfunc`, `defaultIMethod` => `vfuncDefault` (consistent with `VTABLE`, `vptr`) [**BC**].
+ - `externImpl` => `implExtern`, `externDeclImpl` => `declImplExtern`.
  - `self` (`VSelf`) parameters are now `restrict`.
 
 ## [0.7.0] - 2021-10-06